change uint32 to uint16

+++ network/tcp/pack.go
@@ -16,7 +16,7 @@ import (
 )
 
 const (
-	msgLenBytes  uint32 = 2        // 消息长度字节数
+	msgLenBytes  uint16 = 2        // 消息长度字节数
 	msgByteOrder string = "little" // 消息字节排序
 )
 
@@ -29,7 +29,7 @@ func pack(msg []byte) (packet []byte, err error) {
 	var buf bytes.Buffer
 	buf.Grow(len(msg) + int(msgLenBytes))
 
-	if err = binary.Write(&buf, byteOrder(), uint32(len(msg))); err != nil {
+	if err = binary.Write(&buf, byteOrder(), uint16(len(msg))); err != nil {
 		return
 	}
 
@@ -46,7 +46,7 @@ func pack(msg []byte) (packet []byte, err error) {
 func unpack(packet []byte) (msg []byte, err error) {
 	var (
 		buf    = bytes.NewBuffer(packet)
-		msgLen uint32
+		msgLen uint16
 	)
 
 	if err = binary.Read(buf, byteOrder(), &msgLen); err != nil {
@@ -72,7 +72,7 @@ func readMsgFromConn(conn net.Conn, maxMsgLength int) (msg []byte, err error) {
 
 	var (
 		buf    = bytes.NewBuffer(packet)
-		msgLen uint32
+		msgLen uint16
 	)
 
 	if err = binary.Read(buf, byteOrder(), &msgLen); err != nil {
